// src/generated/commentcov/plugin/Block.ts
// Original file: commentcov-proto/commentcov_plugin.proto


export interface Block {
  'startLine'?: (number);
  'startColumn'?: (number);
  'endLine'?: (number);
  'endColumn'?: (number);
}

export interface Block__Output {
  'startLine': (number);
  'startColumn': (number);
  'endLine': (number);
  'endColumn': (number);
}